core: Remove unexpected error and related unnecessary casting (#72)
Unexpected error was a hack to handle items that would otherwise panic
in xdvdfs-web (as it can't catch panics). The lint is failing on nightly
because it's unused if write is not enabled, which is a great reminder
to remove it.

The majority of the unexpected errors are to handle overflow when
casting down to lower sized integer types, so this commit refactors the
code to remove most of these casts in the first place. The remaining ones
are just allowed to panic (and I don't expect it's likely they'll be run
into anyway).