cli: Add benchmark subcommand
Adds a `bmark` subcommand for benchmarking the xdvdfs image creation
loops on real images. For each iteration, the time spent in the forward
and backward passes are measured and reported.

Currently only stdfs and xdvdfs sources are supported as input. In the
future, memoryfs and remapfs sources may be interesting, to remove bias
and to benchmark the remap implementation.

A null blockdevice is added that does not perform any write operations,
eliminating overhead from file copy operations. This enables the
measurement of filesystem read operations and the packing algorithm
itself, which otherwise spends most of its runtime copying files.

The sector linear block device is also supported as an output, which
measures the overhead of copying directory tables into the image. The
copying of files is ideally negligible, but the implementation here
seems inefficient.