core: Improve sector linear filesystem copier performance
Instead of storing each sector separately in the device contents map,
store ranges of data by sector. This removes the loop from copy_file_in,
and greatly reduces the number of entries in the map (proportional to
the number of files/directories in the image, instead of proportional to
the size of the image).